WIP GeoMesa BlobStore Refactoring
Start of refactoring effor to make GeoMesa BlobStore pluggable with both DataStore and BlobStore interfaces.

probably won't build...
