+++ geomesa-blobstore/geomesa-blobstore-core/src/main/java/org/locationtech/geomesa/blob/core/GeoBlobStore.java
@@ -19,7 +19,7 @@ import java.util.Map;
 /**
  * An interface to define how users may ingest and query a GeoMesa BlobStore
  */
-public interface GeoMesaBlobStore extends Closeable {
+public interface GeoBlobStore extends Closeable {
 
     /**
      * Add a File to the blobstore, relying on available FileHandlers to determine ingest
@@ -60,7 +60,7 @@ public interface GeoMesaBlobStore extends Closeable {
 
     /**
      * Deletes a blob from the blobstore by id
-     * @param id id of the blob to delete
+     * @param id id of the blob to deleteBlob
      */
     void delete(String id);
