+++ geomesa-blobstore/geomesa-blobstore-web/src/main/scala/org/locationtech/geomesa/blob/web/BlobstoreServlet.scala
@@ -17,7 +17,7 @@ import javax.servlet.http.{HttpServletRequest, HttpServletRequestWrapper}
 
 import org.apache.commons.io.FilenameUtils
 import org.locationtech.geomesa.accumulo.data.{AccumuloDataStore, AccumuloDataStoreFactory}
-import org.locationtech.geomesa.blob.core.AccumuloBlobStore
+import org.locationtech.geomesa.blob.core.GeoMesaAccumuloBlobStore
 import org.locationtech.geomesa.blob.core.AccumuloBlobStore._
 import org.locationtech.geomesa.utils.cache.FilePersistence
 import org.locationtech.geomesa.web.core.PersistentDataStoreServlet
@@ -50,18 +50,18 @@ class BlobstoreServlet(val persistence: FilePersistence)
       handleError("IO exception in BlobstoreServlet", e)
   }
 
-  val blobStores: concurrent.Map[String, AccumuloBlobStore] = new ConcurrentHashMap[String, AccumuloBlobStore]
+  val blobStores: concurrent.Map[String, GeoMesaAccumuloBlobStore] = new ConcurrentHashMap[String, GeoMesaAccumuloBlobStore]
   getPersistedDataStores.foreach {
     case (alias, params) => connectToBlobStore(params).map(abs => blobStores.putIfAbsent(alias, abs))
   }
 
-  private def connectToBlobStore(dsParams: Map[String, String]): Option[AccumuloBlobStore] = {
+  private def connectToBlobStore(dsParams: Map[String, String]): Option[GeoMesaAccumuloBlobStore] = {
     val ds = new AccumuloDataStoreFactory().createDataStore(dsParams).asInstanceOf[AccumuloDataStore]
     if (ds == null) {
       logger.warn("Bad Connection Params: {}", dsParams)
       None
     } else {
-      Some(new AccumuloBlobStore(ds))
+      Some(new GeoMesaAccumuloBlobStore(ds))
     }
   }
 
@@ -107,7 +107,7 @@ class BlobstoreServlet(val persistence: FilePersistence)
       try {
         persistence.removeAll(persistence.keys(prefix).toSeq)
         persistence.persistAll(toPersist)
-        blobStores.put(alias, new AccumuloBlobStore(ds))
+        blobStores.put(alias, new GeoMesaAccumuloBlobStore(ds))
         Ok()
       } catch {
         case e: Exception => handleError(s"Error persisting data store '$alias':", e)
@@ -163,7 +163,7 @@ class BlobstoreServlet(val persistence: FilePersistence)
       case None => BadRequest(reason = "AccumuloBlobStore is not initialized.")
       case Some(abs) =>
         val id = params("id")
-        logger.debug("Attempting to delete: {} from store: {}", id, alias)
+        logger.debug("Attempting to deleteBlob: {} from store: {}", id, alias)
         try {
           abs.delete(id)
           Ok(reason = s"deleted feature: $id")
@@ -222,7 +222,7 @@ class BlobstoreServlet(val persistence: FilePersistence)
     }
   }
 
-  private def attemptBlobWriting(abs: AccumuloBlobStore, file: FileItem, otherParams: Map[String, String]) = {
+  private def attemptBlobWriting(abs: GeoMesaAccumuloBlobStore, file: FileItem, otherParams: Map[String, String]) = {
     val tempFile = File.createTempFile(UUID.randomUUID().toString, FilenameUtils.getExtension(file.getName))
     try {
       file.write(tempFile)
